move calculateTxSize to common/utils/coin/btc
